requirements.txt
asgiref==3.11.0
Django==4.2.12
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
dj-database-url==2.1.0
djangorestframework-simplejwt==5.2.2
django-tenants==3.5.0
razorpay==1.4.2
